+++ shell.nix
@@ -11,6 +11,10 @@ mkShell {
     # Build toolchain
     rustup
 
+    # Project dependencies
+    pkgconfig
+    openssl
+
     # Other tools
     git
     gitflow